+++ src/main/java/com/example/demo/web/v1/stock/DecreaseStockController.java
@@ -1,8 +1,8 @@
-package com.example.demo.web.v1.product;
+package com.example.demo.web.v1.stock;
 
 import com.example.demo.core.stock.service.DecreaseStockService;
 import com.example.demo.web.ApiResponse;
-import com.example.demo.web.v1.product.request.DecreaseStockRequest;
+import com.example.demo.web.v1.stock.request.DecreaseStockRequest;
 import jakarta.validation.Valid;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
