+++ src/main/java/com/example/demo/web/v1/stock/request/DecreaseStockRequest.java
@@ -1,6 +1,6 @@
-package com.example.demo.web.v1.product.request;
+package com.example.demo.web.v1.stock.request;
 
-import com.example.demo.core.stock.param.IncreaseStockParam;
+import com.example.demo.core.stock.param.DecreaseStockParam;
 import jakarta.validation.constraints.NotEmpty;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -11,7 +11,7 @@ import java.util.stream.Collectors;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class IncreaseStockRequest {
+public class DecreaseStockRequest {
 
     @NotEmpty
     private Set<Stock> stocks;
@@ -27,13 +27,13 @@ public class IncreaseStockRequest {
             this.quantity = quantity;
         }
 
-        public IncreaseStockParam.Stock toParam() {
-            return new IncreaseStockParam.Stock(productCode, quantity);
+        public DecreaseStockParam.Stock toParam() {
+            return new DecreaseStockParam.Stock(productCode, quantity);
         }
     }
 
-    public IncreaseStockParam toParam() {
-        return new IncreaseStockParam(
+    public DecreaseStockParam toParam() {
+        return new DecreaseStockParam(
             stocks.stream().map(Stock::toParam).collect(Collectors.toSet())
         );
     }
